+++ lepo_tests/tests/test_datatypes.py
@@ -3,7 +3,7 @@ from datetime import date, datetime
 import pytest
 from iso8601 import UTC
 
-from lepo.parameters import cast_parameter_value, cast_primitive_value
+from lepo.parameter_utils import cast_parameter_value, cast_primitive_value
 
 DATA_EXAMPLES = [
     {'spec': {'type': 'integer'}, 'input': '5041211', 'output': 5041211},
